requirements.txt
python-dotenv
python-telegram-bot==20.7
markdown
markdownify
beautifulsoup4
